add kafka

+++ src/test/java/com/crudshop/demo/motherobject/CreateProductRequestBuilder.java
@@ -3,6 +3,8 @@ package com.crudshop.demo.motherobject;
 import com.crudshop.demo.controller.product.request.CreateProductRequest;
 import com.crudshop.demo.entity.Categories;
 
+import java.math.BigDecimal;
+
 import static com.crudshop.demo.entity.Categories.FRUIT;
 
 public class CreateProductRequestBuilder {
@@ -10,7 +12,7 @@ public class CreateProductRequestBuilder {
     public static final String DEFAULT_NAME = "Яблоко";
     public static final String DEFAULT_DESCRIPTION = "Круглое красное";
     public static final Categories DEFAULT_CATEGORIES = FRUIT;
-    public static final Double DEFAULT_PRICE = 50.00;
+    public static final BigDecimal DEFAULT_PRICE = BigDecimal.valueOf(50.00);
     public static final Integer DEFAULT_QUANTITY = 15;
     public static final Boolean DEFAULT_IS_AVAILABLE = true;
 
@@ -19,7 +21,7 @@ public class CreateProductRequestBuilder {
     private String name = DEFAULT_NAME;
     private String description = DEFAULT_DESCRIPTION;
     private Categories categories = DEFAULT_CATEGORIES;
-    private double price = DEFAULT_PRICE;
+    private BigDecimal price = DEFAULT_PRICE;
     private Integer quantity = DEFAULT_QUANTITY;
     private Boolean isAvailable = DEFAULT_IS_AVAILABLE;
 
@@ -51,7 +53,7 @@ public class CreateProductRequestBuilder {
         return this;
     }
 
-    public CreateProductRequestBuilder withPrice(Double price) {
+    public CreateProductRequestBuilder withPrice(BigDecimal price) {
         this.price = price;
         return this;
     }
